refactor: create common component for MFA authentication
